Amélioration de l'aspect graphique du site , RETRANSFORMATION des controllers pour reposer sur un model plus objets avec utilisation de $this et non plus de méthode static .

+++ index.php
@@ -4,54 +4,61 @@ define('ROOT', dirname(__FILE__));
 require ROOT . '/controllers/ConnectRegisterController.php';
 require ROOT . '/controllers/ViewFrontendController.php';
 require ROOT . '/controllers/ViewBackendController.php';
+$connectRegisterController = new ConnectRegisterController();
+$viewFrontendController = new ViewFrontendController();
 
 
 if (isset($_GET['action'])) {
     if ($_GET['action'] == 'posts') {
-        ViewFrontendController::showPosts($_GET['nbPage']);
+        $viewFrontendController->showPosts($_GET['nbPage']);
     } elseif ($_GET['action'] == 'post') {
-        ViewFrontendController::showPost($_GET['id']);
+        $viewFrontendController->showPost($_GET['id']);
 //-----------------------------PARTIE CONNECTION--------------------//
     } elseif ($_GET['action'] == 'connect') {
         if (isset($_POST['register'])) {
-            ConnectRegisterController::addUser($_POST['user-name-register'], $_POST['user-pass-register'], $_POST['user-pass-register-verif']);
+            $connectRegisterController->addUser($_POST['user-name-register'], $_POST['user-pass-register'], $_POST['user-pass-register-verif']);
         } elseif (isset($_POST['connect'])) {
-            ConnectRegisterController::connectUser($_POST['user-name'], $_POST['user-pass']);
+            $connectRegisterController->connectUser($_POST['user-name'], $_POST['user-pass']);
         }
-        ViewFrontendController::showConnect();
+        $viewFrontendController->showConnect();
     } elseif ($_GET['action'] == 'deconnect') {
-        ConnectRegisterController::deconnectUser();
-        ViewFrontendController::showAccueil();
+        $connectRegisterController->deconnectUser();
+        $viewFrontendController->showAccueil();
 //-------------------------------------------------------------------//
     } elseif ($_GET['action'] == 'postComment') {
-        ViewFrontendController::showPostAfterPostComment($_GET['id'], $_SESSION['user'], $_POST['commentContent']);
+        $viewFrontendController->showPostAfterPostComment($_GET['id'], $_SESSION['user'], $_POST['commentContent']);
 
     } elseif ($_GET['action'] == 'signaler') {
-        ViewFrontendController::showPostAfterReport($_GET['id'], $_GET['postId']);
+        $viewFrontendController->showPostAfterReport($_GET['id'], $_GET['postId']);
 //-------------------------PARTIE BACKEND-------------------------------//
-    } elseif ($_GET['action'] == 'admin') {
-        ViewBackendController::showBackend($_GET['nbPagePost'], $_GET['nbPageComment']);
-    } elseif ($_GET['action'] == 'addPost') {
-        if (isset($_POST['addPost'])) {
-            echo 'on aajoute un truc ou pas la';
-            ViewBackendController::addPost($_POST['addPostTitle'], $_POST['addPostContent']);
-        } else {
-            ViewBackendController::showAddPost();
-        }
+        //on verifie que l'administrateur du serveur soit bien connecté
+    } elseif (isset($_SESSION['user']) && $_SESSION['user'] == 'admin') {
+        $viewBackendController = new ViewBackendController();
+        if ($_GET['action'] == 'admin') {
+            $viewBackendController->showBackend($_GET['nbPagePost'], $_GET['nbPageComment']);
+        } elseif ($_GET['action'] == 'addPost') {
+            if (isset($_POST['addPost'])) {
+                ViewBackendController::addPost($_POST['addPostTitle'], $_POST['addPostContent']);
+            } else {
+                ViewBackendController::showAddPost();
+            }
 
-    } elseif ($_GET['action'] == 'editPost') {
-        ViewBackendController::showEditPost($_GET['id']);
-    } elseif ($_GET['action'] == 'updatePost') {
-        ViewBackendController::updatePost($_POST['editPostTitle'], $_POST['editPostContent'], $_GET['id']);
-    } elseif ($_GET['action'] == 'deletePost') {
-        ViewBackendController::deletePost($_GET['id']);
-    } elseif ($_GET['action'] == 'editComment') {
-        ViewBackendController::showEditComment($_GET['id'], $_GET['nbPagePost'], $_GET['nbPageComment']);
-    } elseif ($_GET['action'] == 'moderateComment') {
-        ViewBackendController::moderateComment($_GET['id'], $_POST['moderatedComment'], $_GET['nbPagePost'], $_GET['nbPageComment']);
+        } elseif ($_GET['action'] == 'editPost') {
+            ViewBackendController::showEditPost($_GET['id']);
+        } elseif ($_GET['action'] == 'updatePost') {
+            ViewBackendController::updatePost($_POST['editPostTitle'], $_POST['editPostContent'], $_GET['id']);
+        } elseif ($_GET['action'] == 'deletePost') {
+            ViewBackendController::deletePost($_GET['id']);
+        } elseif ($_GET['action'] == 'editComment') {
+            $viewBackendController->showEditComment($_GET['id'], $_GET['nbPagePost'], $_GET['nbPageComment']);
+        } elseif ($_GET['action'] == 'moderateComment') {
+            $viewBackendController->moderateComment($_GET['id'], $_POST['moderatedComment'], $_GET['nbPagePost'], $_GET['nbPageComment']);
+        }
+    } else {
+        echo 'Vous souhaitez acceder à une zone résèrvé à l\'administrateur';
     }
 //---------------------------------------------------------------------------//
 } else {
-    ViewFrontendController::showAccueil();
+    $viewFrontendController->showAccueil();
 }
